refactoring and fixed codacy issue
